requirements.txt
celery==5.1.0
Django==3.2.3
django-allauth==0.44.0 
django-extensions==3.1.3
# flower==0.9.7 # ! dj<3.0
Pillow==8.2.0
psycopg2-binary==2.8.6
gunicorn==20.1.0
python-dotenv==0.17.1
python-memcached==1.59
sorl-thumbnail==12.7.0